Add player controls (#7)
This is the first PR that introduces some notion of a score to the game,
so that's pretty monumental. With this PR, you can catch fish, and you
can sell them. Next, I'd like to tidy up the `update` code as I foresee
it getting uglier, and then... we'll introduce some idle functionality.